Update requirements for Python 3.10.x

+++ utils.py
@@ -14,7 +14,7 @@ def make_cleanup_button(path):
         elif button_text == "Confirm ?":
             for filename in os.listdir(path):
                 file_path = os.path.join(path, filename)
-                if os.path.isfile(file_path) or os.path.islink(file_path):
+                if not filename.startswith(".git") and (os.path.isfile(file_path) or os.path.islink(file_path)):
                     os.unlink(file_path)
                 elif os.path.isdir(file_path):
                     shutil.rmtree(file_path)
